Update kkbox.py
If you want to download all albums of an artist, only the albums shown in https://play.kkbox.com/artist/XXXXXXXXXXXXXX will be downloaded.
But this is only a part of this artist's albums, not all of this artist's albums.
his all albums are showcased at https://play.kkbox.com/artist/DXQVoCQF1asjEu9yaW/albums. and should be downloaded from here
In addition, you seem to use the 'get_date' function in ' rsack\clients\kkbox.py ' to capture the release time of the album through ' Beautifulsoup'. In fact, the release time of the album can also be obtained from 'get_artist_albums'
(In addition, I had the log print out the name of the album currently being downloaded)